# gridlab-studio/frontend-typed/openapi.json
{
  "components": {
    "schemas": {
...
      "GridPreviewBody": {
        "description": "Lightweight request to compute grid rung prices for the live preview.",
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/BacktestRequest"
          }
        },
        "title": "GridPreviewBody",
        "type": "object"
      },
      "GridSearchBody": {
        "properties": {
          "base": {
            "$ref": "#/components/schemas/BacktestRequest"
          },
          "maximize": {
            "default": true,
            "title": "Maximize",
            "type": "boolean"
          },
          "objective": {
            "default": "deflated_sharpe",
            "title": "Objective",
            "type": "string"
          },
          "space": {
            "additionalProperties": {
              "items": {},
              "type": "array"
            },
            "title": "Space",
            "type": "object"
          },
          "top_k": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          }
        },
        "title": "GridSearchBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MonteCarloBody": {
        "properties": {
          "base": {
            "$ref": "#/components/schemas/BacktestRequest"
          },
          "method": {
            "default": "trades",
            "enum": [
              "trades",
              "returns"
            ],
            "title": "Method",
            "type": "string"
          },
          "n_sims": {
            "default": 2000,
            "maximum": 20000.0,
            "minimum": 100.0,
            "title": "N Sims",
            "type": "integer"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          }
        },
        "title": "MonteCarloBody",
        "type": "object"
      },
...
      "RobustnessBody": {
        "properties": {
          "base": {
            "$ref": "#/components/schemas/BacktestRequest"
          },
          "mc_sims": {
            "default": 800,
            "maximum": 10000.0,
            "minimum": 100.0,
            "title": "Mc Sims",
            "type": "integer"
          },
          "n_splits": {
            "default": 3,
            "maximum": 12.0,
            "minimum": 2.0,
            "title": "N Splits",
            "type": "integer"
          },
          "space": {
            "additionalProperties": {
              "items": {},
              "type": "array"
            },
            "title": "Space",
            "type": "object"
          }
        },
        "title": "RobustnessBody",
        "type": "object"
      },
      "RunBacktestBody": {
        "properties": {
          "options": {
            "$ref": "#/components/schemas/BacktestOptions"
          },
          "spec": {
            "$ref": "#/components/schemas/BacktestRequest"
          }
        },
        "title": "RunBacktestBody",
        "type": "object"
      },
...
      "WalkForwardBody": {
        "properties": {
          "base": {
            "$ref": "#/components/schemas/BacktestRequest"
          },
          "n_splits": {
            "default": 4,
            "maximum": 20.0,
            "minimum": 2.0,
            "title": "N Splits",
            "type": "integer"
          },
          "objective": {
            "default": "deflated_sharpe",
            "title": "Objective",
            "type": "string"
          },
          "space": {
            "additionalProperties": {
              "items": {},
              "type": "array"
            },
            "title": "Space",
            "type": "object"
          }
        },
        "title": "WalkForwardBody",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "A beautiful, informative studio for the gridlab grid-trading backtester.",
    "title": "gridlab studio",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/backtest": {
      "post": {
        "operationId": "backtest_api_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBacktestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Backtest Api Backtest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backtest"
      }
    },
    "/api/grid-preview": {
      "post": {
        "operationId": "grid_preview_api_grid_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GridPreviewBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Grid Preview Api Grid Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grid Preview"
      }
    },
...
    "/api/report": {
      "post": {
        "operationId": "report_api_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBacktestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report"
      }
    },
    "/api/research/grid-search": {
      "post": {
        "operationId": "research_grid_search_api_research_grid_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GridSearchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Research Grid Search Api Research Grid Search Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Research Grid Search"
      }
    },
    "/api/research/monte-carlo": {
      "post": {
        "operationId": "research_monte_carlo_api_research_monte_carlo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonteCarloBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Research Monte Carlo Api Research Monte Carlo Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Research Monte Carlo"
      }
    },
    "/api/research/robustness": {
      "post": {
        "operationId": "research_robustness_api_research_robustness_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RobustnessBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Research Robustness Api Research Robustness Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Research Robustness"
      }
    },
    "/api/research/walk-forward": {
      "post": {
        "operationId": "research_walk_forward_api_research_walk_forward_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalkForwardBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Research Walk Forward Api Research Walk Forward Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Research Walk Forward"
      }
    },
    "/api/studio/backtests": {
      "post": {
        "description": "Execute and durably record the typed Studio's migrated Research slice.",
        "operationId": "create_studio_backtest_api_studio_backtests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBacktestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioBacktestRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Studio Backtest"
      }
    },
    "/api/studio/backtests/manifested": {
      "post": {
        "description": "Replay one admitted production dataset without any network dependency.",
        "operationId": "create_manifested_studio_backtest_api_studio_backtests_manifested_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManifestedBacktestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioBacktestRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Manifested Studio Backtest"
      }
    },
    "/api/studio/backtests/{run_id}": {
      "get": {
        "operationId": "get_studio_backtest_api_studio_backtests__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioBacktestRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Studio Backtest"
      }
    },
    "/api/studio/canonical-adaptive": {
      "post": {
        "description": "Present one deterministic legacy-to-canonical adaptive characterization.",
        "operationId": "canonical_adaptive_api_studio_canonical_adaptive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanonicalAdaptiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanonicalAdaptivePresentation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Canonical Adaptive"
      }
    },
    "/api/studio/catalogs/binance/eur": {
      "get": {
        "description": "Return a dated public Testnet/production EUR intersection.",
        "operationId": "get_binance_eur_catalog_api_studio_catalogs_binance_eur_get",
        "parameters": [
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BinanceEurResearchCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Binance Eur Catalog"
      }
    },
...
    "/api/studio/datasets/binance/import": {
      "post": {
        "description": "Download and admit only checksum-verified, continuous production history.",
        "operationId": "import_binance_dataset_api_studio_datasets_binance_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportDatasetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Binance Dataset"
      }
    },
    "/api/studio/datasets/binance/preview": {
      "post": {
        "description": "Resolve official source identities and sizes before archive download.",
        "operationId": "preview_binance_dataset_api_studio_datasets_binance_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BinanceDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BinanceDatasetPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Binance Dataset"
      }
    },
    "/api/studio/datasets/{dataset_id}": {
      "get": {
        "operationId": "get_studio_dataset_api_studio_datasets__dataset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Studio Dataset"
      }
    }
  }
}